+++ burn/Source/KWConverter.h
@@ -9,11 +9,11 @@
 	//movtowav, encodes the movie to wav, after that ffmpeg can encode it
 	NSTask *movtowav;
 	//Status: 0=idle, 1=encoding audio, 2=encoding video
-	int status;
+	NSInteger status;
 	//Number of file encoding
-	int number;
+	NSInteger number;
 	//aspect ratio for current movie
-	int aspectValue;
+	NSInteger aspectValue;
 	//Last encoded file
 	NSString *encodedOutputFile;
 	//Needed for the one who speaks to the class
@@ -22,20 +22,20 @@
 	BOOL userCanceled;
 	
 	//Input file values
-	int inputWidth;
-	int inputHeight;
+	NSInteger inputWidth;
+	NSInteger inputHeight;
 	float inputFps;
-	int inputTotalTime;
+	NSInteger inputTotalTime;
 	float inputAspect;
 	//inputFormat: 0 = normal; 1 = dv; 2 = mpeg2
-	int inputFormat;
+	NSInteger inputFormat;
 
 	NSDictionary *convertOptions;
 	NSString *errorString;
 	NSString *convertDestination;
 	NSString *convertExtension;
-	int convertRegion;
-	int convertKind;
+	NSInteger convertRegion;
+	NSInteger convertKind;
 	
 	BOOL useWav;
 	BOOL useQuickTime;
@@ -45,19 +45,19 @@
 //Encode actions
 
 //Convert a bunch of files with ffmpeg/movtoyuv/QuickTime
-- (int)batchConvert:(NSArray *)files withOptions:(NSDictionary *)options errorString:(NSString **)error;
+- (NSInteger)batchConvert:(NSArray *)files withOptions:(NSDictionary *)options errorString:(NSString **)error;
 //Encode the file, use wav file if quicktime created it, use pipe (from movtoy4m)
-- (int)encodeFileAtPath:(NSString *)path;
+- (NSInteger)encodeFileAtPath:(NSString *)path;
 //Encode sound to wav
-- (int)encodeAudioAtPath:(NSString *)path;
+- (NSInteger)encodeAudioAtPath:(NSString *)path;
 //Stop encoding (stop ffmpeg, movtowav and movtoy4m if they're running
 - (void)cancelEncoding;
 
 //Test actions
 
 //Test if ffmpeg can encode, sound and/or video, and if it does have any sound
-- (int)testFile:(NSString *)path;
-//Test methods used in (int)testFile....
+- (NSInteger)testFile:(NSString *)path;
+//Test methods used in (NSInteger)testFile....
 - (BOOL)streamWorksOfKind:(NSString *)kind inOutput:(NSString *)output;
 - (BOOL)isReferenceMovie:(NSString *)output;
 - (BOOL)setTimeAndAspectFromOutputString:(NSString *)output fromFile:(NSString *)file;
@@ -80,14 +80,14 @@
 - (NSArray *)succesArray;
 
 //Other actions
-- (int)convertToEven:(NSString *)number;
-- (int)getPadSize:(float)size withAspect:(NSSize)aspect withTopBars:(BOOL)topBars;
+- (NSInteger)convertToEven:(NSString *)number;
+- (NSInteger)getPadSize:(float)size withAspect:(NSSize)aspect withTopBars:(BOOL)topBars;
 - (BOOL)remuxMPEG2File:(NSString *)path outPath:(NSString *)outFile;
 - (BOOL)canCombineStreams:(NSString *)path;
 - (BOOL)combineStreams:(NSString *)path atOutputPath:(NSString *)outputPath;
-- (int)totalTimeInSeconds:(NSString *)path;
+- (NSInteger)totalTimeInSeconds:(NSString *)path;
 - (NSString *)mediaTimeString:(NSString *)path;
-- (NSImage *)getImageAtPath:(NSString *)path atTime:(int)time isWideScreen:(BOOL)wide;
+- (NSImage *)getImageAtPath:(NSString *)path atTime:(NSInteger)time isWideScreen:(BOOL)wide;
 - (void)setErrorStringWithString:(NSString *)string;
 
 @end
